Add `#new_record?` to save operations
`#persisted?` always returns `true` in `after_save` and
`after_commit`, so it is not a viable method for checking
whether or not we did a create (vs. update) operation.

`#new_record` will return `true` if this is a new record we are
creating (or we just created), and `false` otherwise, independent
of which callback we are in.