Ensure password is always saved when resetting password
`SavePassword` operation does not save password if the new
password is same as the old. This is useful for ensuring the
password hash is not changed, thus notification is not sent.

In `ResetPassword`, we need the hash to change even if the new
password is same as old, in order to trigger `after_save` and
delete the reset token.

We should probably find a better solution, though.