Rewrite bearer logins
We're separating API actions from browser actions, and replacing
session-based authentication with temporary tokens for APIs,
among others.

Using sessions for APIs is unreliable, due to same origin policy
and CORS, especially when the API consumer is a browser-based app.

Further, to strengthen security for session authentication, we
may have had to include CSRF protection for APIs, which may
require sending back a CSRF token for almost every action
performed by the API consumer.

To wit, it's not worth the trouble.