Fix db connection error when db renamed
This was caused by the compose config file living in the `spec/`
directory.

It made the default compose project name `spec`, which resulted in
namespace clashes with other projects on my local machine, which
had their compose config file in their respective `spec/`
directories.

Now our command are shorter, cuz we do not have to specify the
config file path.