Add login idle timeout
This logs a user out if they have not engaged with the application
for a given period.

We are not applying this to APIs, at least, for now. Because this
implementation is session-based, and we're not using sessions in
authentication; session cookies are not guaranteed to be sent back
from clients in APIs, due to same origin policy.