Fix `before_save` callbacks called multiple times in a single save
Fixes https://github.com/GrottoPress/shield/issues/6

Finally got the specs to pass with `after_completed` introduced in
*Avram* v0.18.0.

`after_completed` seems to complicate the API, and does not help
in situations where you need to roll back a parent operation's
transaction if a nested operation fails; it would be too late to
roll back.

In such cases, I ended up calling the same method in `after_save`
and `after_completed`, and checking if no attributes changed (ie,
if `after_save` did not run) in the latter.