Add more specs
Aiming for complete test coverage. This is particularly important
because *Crystal*'s semantic analyzer is lazy, and would not check
a method or file that is not called or used.

This means you could end up with a compile error without warning,
if that block of code leading to the error is never called.