Add presets
Presets sets up most things for you. You do not have to go through
the cumbersome process of setting up every type.

The following are **not** included. They should be set up manually:

- Models (except `BaseModel`)
- Migrations
- Actions (except `BrowserAction` and `ApiAction`)

This is mostly because these are optional, and every application
should determine which of each it needs.

Of course, you still need to set up your emails, serializers
and pages.