Remove `Avram::Box.has_one` macro
It was a nice experiment, but I realised one may not always want
to create a nested box when a box is created.

The nature of boxes is such that you cannot create more than one
box per model, unlike operations. So you cannot, for instance,
create a user box that has one user options, and another user box
that doesn't.

So this macro created coupling in a way that could make testing
inflexible.